Rename module 'models.py' as 'dspace.py'
